+++ static/pong/js/bot.mjs
@@ -6,7 +6,7 @@
  * Date: 15/10/2023
  * author: BoxBoxJason
  */
-import { paddle2, game_frame, ball } from './constants.mjs'
+import { paddle2, game_frame, ball, paddle_height, ball_size } from './constants.mjs'
 
 // Bot difficulty level
 var bot_level = 1;
@@ -16,17 +16,19 @@ var bot_level = 1;
 document.getElementsByName('difficulty').forEach(button => {
 	button.addEventListener('change', (e) => {
 		bot_level = parseInt(button.value);
+        console.log(bot_level);
 	});
 });
 
 
 export function moveBotPaddle() {
 	// Moves bot paddle according to chosen difficulty
-    if (ball != null && game_frame != null && paddle2 != null){
-        let gameFramePos = game_frame.getBoundingClientRect();
-        let newTopCoord = window.pageYOffset + ball.getBoundingClientRect().top - 75;
+    if (ball != null && game_frame != null && paddle2 != null) {
+        const gameframe_bounding_rect = game_frame.getBoundingClientRect();
+        const ball_bounding_rect = ball.getBoundingClientRect();
+        let newTopCoord = window.pageYOffset + (ball_bounding_rect.top + ball_bounding_rect.bottom) / 2 - paddle_height - ball_size;
         newTopCoord = Math.max(0,newTopCoord);
-        newTopCoord = Math.min(newTopCoord,gameFramePos.height - 100);
+        newTopCoord = Math.min(newTopCoord,gameframe_bounding_rect.height - 100);
 
         setTimeout(() => paddle2.style.top = newTopCoord + 'px',200 - 15 * bot_level);
     }