+++ static/pong/js/constants.mjs
@@ -11,8 +11,9 @@ export const game_frame = document.getElementById('gameFrame');
 // Paddles
 export const paddle1 = document.getElementById("paddle1");
 export const paddle2 = document.getElementById("paddle2");
+export const paddle_height = 100;
 // Ball
 export const ball = document.getElementById("ball");
-
+export const ball_size = 30;
 // Number of rebounds before speed increase
 export const number_rebounds_speed_increase = 7;