+++ static/resources/js/utils.mjs
@@ -1,0 +1,12 @@
+/* JavaScript FRONT
+ * Project: GameBox
+ * Package: resources
+ * Module: utils
+ * Usage: General utility functions.
+ * Date: 23/12/2023
+ * author: BoxBoxJason
+ */
+
+export function basename(path) {
+    return path.split('/').pop();
+}